+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `StableVec::contains()`
 - `StableVec::into_vec()`
 - `impl<T> Default for StableVec<T>`
+- Added `FromIterator` impl for `StableVec`
 - Added `Debug` impl for `StableVec` with a fitting example